Running 'make clean-all' has errors if a dep has no Makefile
Create a project with a dependency that has no Makefile such as wooga/eredis.
Do 'make all' then 'make clean-all' to get an error. This patch checks for the
presence of the Makefile before attempting to run 'make clean', however, it does
not attempt run 'rebar clean' or other commands instead.