Wrapped assignment in an ifeq.
This change is to wrap this setting for ```ERL_LIBS``` in a conditional
so the assignment is more carefully controlled.

If ```ERL_LIBS``` is undefined or empty then the original intent will
continue unaffected and ```DEPS_DIR``` will be assigned as it was
before.

Otherwise, the existing value of ```ERL_LIBS``` will be read in and the
value of ```DEPS_DIR``` will be appended to it using the ```addsuffix```
command.

From what I understand this is a more robust method of building the
variable within a Makefile.