Change package index file format and deps handling
This commit adds two columns to the package index file: the
method used to retrieve the dependency (only git supported
at this point) and the preferred version (a commit or
equivalent in the case of git).

Now that all the necessary information is in the package
index file, the dep_* lines become optional. It is possible
to fetch dependencies by just listing it in the DEPS variable,
for example this would fetch both cowboy and jsx:

DEPS = cowboy jsx

The dep_* line can be used to override the defaults. It now
takes three values, the method used to retrieve the dependency,
the repository URL and the preferred version. The only change
from before is the first value being added.