Fail if no empty modules entry in .app.src
An empty list is required for the modules tuple, as in
{modules, []}
so a sed call could populate the list. This is mentioned in README, but
can be overlooked.

If the empty tuple is not present, relx will fail on missing dependency
errors. Therefore, fail make if the empty tuple isn't found, to
help the user diagnose the problem.