Support dependencies fetched as git submodules
This adds an alternative argument, 'submodule', as an alternative to a
URL for git dependencies. For example:

dep_cowboy = git submodule

This option is not propagated to sub-dependencies and all required
submodules should be added to the root repository.

The main benefit is that in later version of git, all references are
stored in the main repositories directory (i.e. .git/modules). This
allows for speedy clean and build without remote cloning every time.