Introduce test builds and unify testing tools interface
The general idea is that erlang.mk now keeps track of what kind
of build it generated. A test build is valid for all subsequent
test target invocations. A normal build is only valid for itself
and releases.

This rework adds the ability to specify deps to eunit.

The EUNIT_DIR variable is gone in favor of a more global TEST_DIR.

The tests-ct target got renamed to ct and documented.

Many more minor changes were done during the course of testing
these changes.