Add "necessary evil" autopatch feature
The point of this feature is to fix common issues users
experience.

This feature, when used, attempts to fix projects that
are incompatible with erlang.mk. It currently is able
to build a new Makefile based on information it finds
in a rebar.config file (only deps for now) and also
to fix the .app.src file in various ways.