Initial rebar2 plugins support
We do nothing more than what is necessary to make riak_pb work.
It's possible other plugins will not work, however this is the
only plugin that I have seen to be required so far, and hook
plugins seem to have been removed from rebar3, so it should be
fine.