Compile dependencies only once
We keep track of which dependencies we compiled in the current
session in the $(ERLANG_MK_TMP)/deps.log file.

This will help save a little time when projects depend more than
once on the same dependency. While going in the directory and
running make was already quick, it's quicker if not, and the
logs end up cleaner.