Add RABBITMQ_CLIENT_PATCH to make upstream amqp_client work
While this is not quite "just work" due to other packages
depending on a forked off rabbitmq, this is still pretty
good as it actually makes it work and still leaves a choice
between upstream and the fork.

Hopefully people will start to use upstream but you know how
these things tend to go.