core/deps-tools.mk: New targets `fetch-deps` and `list-deps`
.. to fetch and list deps recurvively.

Therefore, they work on direct dependencies, dependencies' dependencies,
and so on. Nothing is built with those targets.

The following targets are also available to work on other kinds of
dependencies:
    make fetch-doc-deps
    make fetch-rel-deps
    make fetch-test-deps
    make fetch-shell-deps
    make list-doc-deps
    make list-rel-deps
    make list-test-deps
    make list-shell-deps

In all cases, they include "normal" and eg. "test" dependencies of the
top-level project, then only "normal" dependencies' dependencies.

It's possible to include several kinds in one go:
    make fetch-deps DEP_TYPES='doc test'
    make list-deps DEP_TYPES='doc test'

As it may be difficult to use the output of `make list-*` because
the list could appear after other targets output (like dependencies
fetching), the list is available in files pointed by the following
variables:
    $(ERLANG_MK_RECURSIVE_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_DOC_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_REL_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_TEST_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_SHELL_DEPS_LIST)

Internally, `list-deps` is simply implemented on top of `fetch-deps`:
the latter fills a sorted log with all the directories it traversed.
`list-deps` finally just prints the log file to stdout.

Fixes #560.