Add initial 'make rebar.config'
This is a preliminary work, good enough to be used in some,
but not all, projects. Tests cover the current functionality.
Documentation has been written but will be committed at a
later stage, as it covers more than just this target (the
target is documented as part of a "Compatibility with other
build tools" chapter).