Make tests work on more systems
Some systems do not have subsecond file modification detection
capabilities. This is the case in OSX because the HFS+ file
system has 1 second resolution. This is also the case in other
OSes with particular file systems or even shells. For example,
the "test" command in Bash does not support subsecond file times,
while the ZSH one and the /usr/bin/test on my machine does. On
those systems, with Bash, find -newer will work, but not test -nt.

Tests were updated to reflect this. The test Makefile runs a
series of commands to determine if subsecond resolution is available.
When it is not, a sleep command will be executed before relevant
operations in the tests.

Because the Bash shell is often the default shell for Make, and
because users of ZSH may want to avoid sleeping for no reasons,
a new variable was introduced: ZSH=1. When provided when running
tests, the shell will be switched to ZSH, and subsecond resolution
will work as expected. Tests have been fixed to work with ZSH as
well.

All this only applies to tests, Erlang.mk itself is so far not
affected by this issue.