erlc.mk: Handle .app generation outside of .erl compilation
Previously, if no source file were changed, the `ebin/$(PROJECT).app`
rule was not evaluated again. Therefore, the `.app` file was not
updated, even though, `$(GITDESCRIBE)` or `.app.src` were changed.

Now, the `.app` file is handled separately and always evaluated. Before
overwriting the existing `.app` file though, we check if the content of
the new file is different.

In the testsuite, the `.app` file is now unmodified.