Use `:` as a noop instead of `echo -n`
The `-n` flag is not standard and not portable. It is supported by many
Bourne shell to suppress the trailing newline character. But some old
shells do not support this and "-n" is printed.

This is the case for instance on some old versions of Mac OS X where
`/bin/sh` is Bash 3.2.