Use an intermediate timestamp file to know when to touch source files
The previous approach spawned a shell for every single source files. As
we disable parallel make jobs, this is very time consuming.

The new approach uses an intermediate timestamp file in $(ERLANG_MK_TMP)
to record the last modification of any of $(MAKEFILE_LIST) and touch
all source files in one command. Then, the .app file depends on this
timestamp file.

We test the existence of this timestamp file: if if doesn't exist, don't
touch source files, they will be built anyway.

$(PROJECT).d now depends directly on $(MAKEFILE_LIST); before, this
dependency was indirect through $(ERL_FILES). Also, once $(ERL_FILES)
were touched, we do the same with $(PROJECT).d because there is no need
to regen it because of this.