Remove dhtcrawler from the package index
The URIs don't work for non-logged in users.

Choosing to remove for now because the project also changed
(there's a dhtcrawler2) so it might be worth adding this one
instead (although it probably won't be compatible with Erlang.mk
right now and seems to be more of a standalone repository).