Temporarily remove restc and oauth2c
The restc project has dependencies that require to be logged in
to fetch them, and this breaks the packages test suite. The oauth2c
project depends on restc.

I sent a pull request at https://github.com/kivra/restclient/pull/8

The packages can be added back just fine after this pull request
get merged.

In the future it might be worth having a value marking packages
as broken so they are skipped by the test suite.