Fix compilation of NIFs on Windows
Thanks to two users of the ninenines/esdl2 project,
a correct way to build NIFs on Windows has been found.

At the moment we require a specific compiler (MingW's
gcc). Maybe we can change this in the future and allow
Visual Studio and others.

Some small changes have been made to the documentation,
and the meaning of one configuration variable has changed
to not include the extension (which is decided automatically
by Erlang.mk; and configurable separately).

Enjoy!