Use curl instead of wget/erl for fetching files
After a lot of soul searching and experiments, I am opting
to using curl for fetching files. It works with very little
manipulation on all supported systems. Some might need it
installed, but most come with it.

While wget was working, it had TLS validation deactivated
because some users didn't have CA certificates on their
system or wget was not configured properly. This does not
seem to be a problem with curl as far as I tested, in
particular on OSX.

The fallback to use Erlang was also insecure, and probably
not the best idea. Let's keep things simple. Curl is required.