Add self-extracting/running archives for releases
This is a proof of concept at the moment. It only runs
the release in 'console' mode at the moment. I am expecting
users to want a mix of both self extracting (without removing
when done) and self running (with removal when done) archives.
There is currently no way to configure the behavior and so
the files get deleted when the release terminates (self running
use case).

The SFX archive will only be created when the SFX variable
is set (either in the Makefile or through the command-line).

There are currently no documentation. Just "make SFX=1".