Only makedep and build asn1/xrl/mib files when necessary
Because we were unconditionally including the makedep file,
it would be built even if the target was unrelated (like
"make help"), and it would also trigger building of xrl
and other files that produce Erlang code.

Instead the makedep file is now included as a wildcard:
if it exists it gets included, if it doesn't it's not,
and if it's created in the middle of execution it gets
included only then.

More details on the technique can be found at

  http://make.mad-scientist.net/papers/advanced-auto-dependency-generation/

A test has been added to make sure the behavior does not
come back at a later time.