Make sure the build fails if a docs/rel/test/shell dep fails
Before this change, the build would continue, even if a dependency
failed to build. This could lead to obscure errors in the middle of a
testsuite for instance.

With this change, the build fails immediately, exactly like when a
regular dependency fails to build.

While here, replace most uses of `|| exit $$?` with `set -e`. This
simplifies error handling if we need to add more commands to each
blocks.

Also, echo error messages to stderr.