Add "make install p=$PKG" to install packages locally
This uses the erlang.mk test infrastructure to build the
package and the packages are installed by default to
XDG_DATA_HOME/erlang.mk/lib. Just add this path to
ERL_LIBS and it will be available to all non-release
Erlang shells.

  export ERL_LIBS=$HOME/.local/share/erlang.mk/lib

To make them available to releases they need to be
started in interactive mode. This can be done via
the CODE_LOADING_MODE variable which is used at the
release startup.

  export CODE_LOADING_MODE=interactive