Fix jobserver unavailable warnings
When the $(MAKE) is inside an $(eval $(call ...)) then it needs
to be escaped as $$(MAKE).

When the $(MAKE) is inside a $(call ...) then I did not figure
out a way other than passing it as an argument.

When the $(MAKE) is inside many levels of $(call $(call ...))
it's easier to avoid it if at all possible, so I replaced the
rebar $(MAKE) call with ./bootstrap. I confirmed it works fine
on Windows as well.