Fix rebuild on Makefile change when new deps are added
The issue was that when a new dep is added the last-makefile-change
update was done in the dependency and not in the top-level Makefile,
preventing the rebuild of the top-level project.

A test was also added to do the same when editing dependencies and
things seem to work properly in that case.

Thanks to robinchew on IRC for the great help debugging this.