Fix concurrent creation of `recursive-*deps-list.log` files
To do that, we simply include the PID of the Make process in temporary
file.

We also compare the finished temporary file with the actual target file.
If they are different, the temporary file is moved to the final target,
otherwise, we just get rid of the temporary file. This way, the target
timestamp stays the same if the content does not change.