Account for different ways of defining deps
Rather than looking for ^dep_* variables in a few locations which are
guaranteed to miss some definitions, use the deps dirs that Erlang.mk
resolves recursively, then start listing deps with info (renamed from
show-deps) until the target is either not implemented (maybe erlang.mk
needs updating) or there are no more deps defined.

FWIW, Erlang.mk injects the original context in the direct deps
(a.k.a. "the deps of the first deps", a.k.a. "first-level deps"), but it
does not carry the initial context past the first level. This means that
deps of deps of deps (i.e. rabbitmq-common which is a a dep of
rabbitmq-management which is a dep of the current project) need to
implement list-deps-info + %-list-deps-info for the recursion to
continue (captured in one of the TODOs).

Pair @dumbbell

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>