Add the query-deps family of targets
The query-deps, query-doc-deps, query-rel-deps, query-test-deps
and query-shell-deps targets go through dependencies and print
some information about them:

  $ make query-deps
  ...
  cowboy: cowlib git https://github.com/ninenines/cowlib 2.7.3
  cowboy: ranch git https://github.com/ninenines/ranch 1.7.1

  $ make query-deps QUERY="name repo version absolute_path"
  ...
  cowboy: cowlib https://github.com/ninenines/cowlib 2.7.3 /path/to/deps/cowlib
  cowboy: ranch https://github.com/ninenines/ranch 1.7.1 /path/to/deps/ranch

The query-deps target is recursive; the others aren't. This
mirrors the behavior when fetching and building dependencies.

The full list of options is:

  QUERY="fetch_method name repo version extra absolute_path"

When an option has no corresponding value, "-" will be printed.
For example most fetch methods do not have extra information to
print. When there is extra information, the value will be prefixed
with a descriptive name. For example: package-name=uuid_erl

This commit also introduces changes to deps handling: some of
the functions for querying will now be used directly. In the
future the goal is to remove the old dep_name, dep_commit, etc.
functions and replace their usage with the new query functions.

Custom fetch methods should implement query functions in order
to have information about the relevant dependencies printed
properly.