plugin_dialyzer.mk: Update erlang.mk in `deps/cowlib`
The problem is that the copy of `erlang.mk` in cowlib does not have
locking in `dep_autopatch_fetch_rebar`. This triggers failures in
parallel testing of Erlang.mk.

A second improvement is to just test-build cowlib instead of actually
running the tests. Test deps are fetched and compiled, but we don't
waste time running cowlib's testsuite.