Fix encoding issues when generating the makedep file
In some cases the input will be parsed as UTF-8 and converted
to "characters" and in others it won't and will be processed
as bytes (for example, `erl -oldshell` will do that). This means
that encoding-dependent characters such as "é" only need to be
converted to binary in the former case.

To detect which situation we are in we check what the value of
"é" is. If it is [233] then the eval input was parsed as UTF-8
and converted to characters. Otherwise we assume it wasn't.