Remove packages explicitly unmaintained or deleted
The packages either no longer have a project page on GitHub
(in some cases the users were deleted) or the users have
archived the projects and they no longer compile.

The packages from jj1bdx removed in this commit were
integrated in the 'rand' module found in the OTP distribution.