Remove most packages
Erlang.mk will no longer feature packages out of the box,
except for the applications that we implement plugins for
(such as erlydtl, proper or relx) and the projects I work
on that are also used to test Erlang.mk (cowboy and friends).

This is a breaking change. In most cases the fix is to add
a full dep_* line for the dependencies that were used as
packages before.