test/core_elixir.mk: Specify CFLAGS when building the Elixir NIF
On FreeBSD, libsodium's `sodium.h` header is installed in
`/usr/local/local`. The Makefile already adds
`/usr/local/include/sodium` to the compiler's `-I` search path, but it
doesn't cover the FreeBSD case.

This patch adds `-I /usr/local/include` to the CFLAGS on the make(1)
command line to let libsalty2's Makefile pick it up. This fixes the
build on FreeBSD.