GitHub Actions: Run tests on FreeBSD
Tests are executed in a FreeBSD VM using the latest stable release. The
default version of Erlang and Elixir from the Ports tree are being used;
currently Erlang/OTP 26 and Elixir 1.17 as of this commit.

Otherwise, tests are split the same way as for Linux/Windows/OSX. That
said, the list is duplicated, so a bit annoying to maintain.