Avoid calling `core_native_path` in dep_fetch_hex many times
On Windows, this executes a shell to determine the native path, which is
very expensive. The result is always the same, so make sure
`core_native_path` is called once instead.

For example, this saves a significant amount of time when compiling
RabbitMQ on Windows.

LH: Moved the native variables near hex_get_tarball.erl since
    they're only needed there.