GitHub Actions: Set a timeout of 1 hour for `check-in-vm`
It looks like the VM is sometimes stuck booting, long before executing
any tests. The default timeout of 6 hours blocks other CI jobs for a
very long time. Because the VM won't boot anyway, we reduce this time to
one hour.

One hour should be long enough for a successful slow execution of the
tests.