GitHub Actions: Switch to `cross-platform-actions` actions to run VMs
`vmactions` VM fails to boot quite frequently and we have no visibility
on the reason of the failure.

The only failure I got out-of-the-box with cross-platform-actions VM was
the `core-deps-rel` test case which hangs. It comes from the use of
`erl_call` to detect if the Erlang node was started correctly. A comment
in the generated release starts scripts says:

    # users who depend on stdout when running rpc calls must still use nodetool
    # so we have an overload option to force use of nodetool instead of erl_call

If `$USE_NODETOOL` is set, it will use `relx_nodetool` instead. Indeed,
this fixed the hang.